move interface, type definitions to types file
